support `16-true` and test it instead of `16-mixed` with FSDP

+++ src/fts_examples/legacy/ipynb_src/fts_superglue_nb.py
@@ -579,7 +579,7 @@ def train() -> None:
         enable_progress_bar=enable_progress_bar,
         max_epochs=1,
         limit_train_batches=2,
-        precision="16-mixed",
+        precision="16-true",
         accelerator="auto",
         devices=1 if is_cuda_available() else None,
         callbacks=callbacks,
